requirements.txt
python-telegram-bot==20.8
appwrite==4.0.0
python-dotenv==1.0.0
